switch hook to be a controller
